requirements.txt
dbus-python==1.2.12
pypresence==3.3.2
